fix: keep mobile Enter as newline and improve composer hint spacing
- Mobile: Enter no longer submits the prompt (line break instead); the
  send button is the only way to send. Desktop keeps Enter = submit and
  Shift+Enter = newline.
- Add top spacing to the "Agent working - send still queues" hint so it
  no longer sits flush against the input bar.
- Extract a shared useIsMobile() hook (matching the existing
  max-width:767px definition) and consume it from use-collapsible-toolbar
  to remove duplicated matchMedia logic.
